+++ src/main/java/net/atos/client/zgw/zrc/ZrcClientService.java
@@ -7,7 +7,7 @@ package net.atos.client.zgw.zrc;
 
 import static java.lang.String.format;
 import static net.atos.zac.configuratie.ConfiguratieService.ENV_VAR_ZGW_API_CLIENT_MP_REST_URL;
-import static net.atos.zac.util.UriUtilKt.uuidFromURI;
+import static net.atos.zac.util.UriUtilsKt.uuidFromURI;
 
 import java.net.URI;
 import java.util.Collection;
