+++ src/main/java/net/atos/zac/app/inboxdocumenten/InboxDocumentenRESTService.java
@@ -5,7 +5,7 @@
 
 package net.atos.zac.app.inboxdocumenten;
 
-import static net.atos.zac.util.UriUtilKt.uuidFromURI;
+import static net.atos.zac.util.UriUtilsKt.uuidFromURI;
 
 import java.util.List;
 import java.util.Optional;
