+++ src/main/java/net/atos/zac/app/informatieobjecten/converter/RestInformatieobjectConverter.java
@@ -10,7 +10,7 @@ import static net.atos.zac.app.configuratie.model.RestTaalKt.toRestTaal;
 import static net.atos.zac.app.identity.model.RestUserKt.toRestUser;
 import static net.atos.zac.configuratie.ConfiguratieService.OMSCHRIJVING_TAAK_DOCUMENT;
 import static net.atos.zac.identity.model.UserKt.getFullName;
-import static net.atos.zac.util.UriUtilKt.uuidFromURI;
+import static net.atos.zac.util.UriUtilsKt.uuidFromURI;
 import static nl.lifely.zac.util.Base64ConvertersKt.toBase64String;
 
 import java.time.LocalDate;
