+++ src/main/java/net/atos/zac/documenten/InboxDocumentenService.java
@@ -5,7 +5,7 @@
 
 package net.atos.zac.documenten;
 
-import static net.atos.zac.util.UriUtilKt.uuidFromURI;
+import static net.atos.zac.util.UriUtilsKt.uuidFromURI;
 
 import java.time.ZoneId;
 import java.util.ArrayList;
